fix(editor): Key uint64 debugger dictionaries with matching carrier
TypedDictionary<uint64_t, Variant> and TypedArray<uint64_t> now type-check
their key/element carrier as Variant::UINT, since native uint64_t declares
that carrier. EditorDebuggerInspector inserted raw ObjectID values (which
keep the signed Variant::INT carrier) and a bare int literal 0 into these
containers without the uint64_t cast used elsewhere in the same file,
so the insert silently failed validation and left the container without
that key.

The dropped keys made the remote-object property cache spin on repeated
"Dictionary::operator[] used when there was no value" errors after a DAP
restart's second breakpoint stop, starving the debug adapter thread and
timing out the client's "continue" request. Cast to uint64_t at each site,
matching the existing pattern already used for array insertions in the
same functions.

Also route the restart "continue"/"terminated" assertions in
test_editor_tooling_host.h through the existing tooling-host-and-DAP
diagnostic helper so a future failure carries the debuggee log and DAP
message trace instead of a bare message.

+++ editor/debugger/editor_debugger_inspector.cpp
@@ -214,7 +214,7 @@ EditorDebuggerRemoteObjects *EditorDebuggerInspector::set_objects(const Array &p
 				UsageData usage_dt;
 				usage_dt.prop = prop;
 				usage_dt.prop.first.name = pinfo.name;
-				usage_dt.values[obj.id] = prop.second;
+				usage_dt.values[(uint64_t)obj.id] = prop.second;
 				usage[pinfo.name] = usage_dt;
 			}
 
@@ -240,7 +240,7 @@ EditorDebuggerRemoteObjects *EditorDebuggerInspector::set_objects(const Array &p
 				}
 
 				usage[pinfo.name].qty++;
-				usage[pinfo.name].values[obj.id] = prop.second;
+				usage[pinfo.name].values[(uint64_t)obj.id] = prop.second;
 			}
 		}
 
@@ -462,7 +462,7 @@ void EditorDebuggerInspector::add_stack_variable(const Array &p_array, int p_off
 		}
 		variables->prop_list.insert_before(current, pinfo);
 	}
-	variables->prop_values[pinfo.name][0] = v;
+	variables->prop_values[pinfo.name][uint64_t(0)] = v;
 	variables->update();
 	edit(variables);
 }
